requirements.txt
flask
python-dotenv
flask-pymongo
flask-login
openai
langchain_openai
langchain